# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Meine Bibliothek",
    "description": "ÜK 295",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/books": {
      "get": {
        "tags": [
          "Book"
        ],
        "summary": "Get all books",
        "description": "Get all books currently in the library",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Book"
        ],
        "summary": "Create a new book",
        "description": "Create a new book in the library",
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/books/{isbn}": {
      "get": {
        "tags": [
          "Book"
        ],
        "summary": "Get a book by isbn",
        "description": "Get a book by isbn from the library",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Book"
        ],
        "summary": "Update a book",
        "description": "Update a book in the library",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "delete": {
        "tags": [
          "Book"
        ],
        "summary": "Delete a book",
        "description": "Delete a book from the library",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Book"
        ],
        "summary": "Update properties of a book",
        "description": "Update properties of a book in the library",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/lends": {
      "get": {
        "tags": [
          "Lend"
        ],
        "summary": "Get all lends",
        "description": "Get all lends currently in the library",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Lend"
        ],
        "summary": "Create a new lend",
        "description": "Create a new lend in the library",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "isbn": {
                  "example": "any"
                },
                "customer_id": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                },
                "borrowed_at": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/lends/{id}": {
      "get": {
        "tags": [
          "Lend"
        ],
        "summary": "Get a lend by id",
        "description": "Get a lend by id from the library",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Lend"
        ],
        "summary": "Return a lend",
        "description": "Return a lend to the library",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Book": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "isbn": {
          "type": "string",
          "example": ""
        },
        "title": {
          "type": "string",
          "example": ""
        },
        "author": {
          "type": "string",
          "example": ""
        },
        "year": {
          "type": "number",
          "example": 0
        }
      },
      "required": [
        "isbn",
        "title",
        "author",
        "year"
      ]
    },
    "Lend": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "customer_id": {
          "type": "string",
          "example": ""
        },
        "isbn": {
          "type": "string",
          "example": ""
        },
        "borrowed_at": {
          "type": "string",
          "example": ""
        },
        "returned_at": {
          "type": "string",
          "example": ""
        }
      },
      "required": [
        "customer_id",
        "isbn",
        "borrowed_at",
        "returned_at"
      ]
    }
  }
}